+++ nevergrad/benchmark/experiments.py
@@ -3,18 +3,20 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
+import typing as tp
 from typing import Iterator, Optional, List, Union, Any
 import numpy as np
 import nevergrad as ng
-from ..functions import ExperimentFunction
-from ..functions import ArtificialFunction
-from ..functions import MultiobjectiveFunction
-from ..functions import mlda as _mlda
-from ..functions.arcoating import ARCoating
-from ..functions.powersystems import PowerSystem
-from ..functions.stsp import STSP
-from ..functions import rl
-from ..functions.games import game
+from nevergrad.functions import ExperimentFunction
+from nevergrad.functions import ArtificialFunction
+from nevergrad.functions import FarOptimumFunction
+from nevergrad.functions import MultiobjectiveFunction
+from nevergrad.functions import mlda as _mlda
+from nevergrad.functions.arcoating import ARCoating
+from nevergrad.functions.powersystems import PowerSystem
+from nevergrad.functions.stsp import STSP
+from nevergrad.functions import rl
+from nevergrad.functions.games import game
 from .xpbase import Experiment as Experiment
 from .xpbase import create_seed_generator
 from .xpbase import registry as registry  # noqa
@@ -191,6 +193,7 @@ def paramultimodal(seed: Optional[int] = None) -> Iterator[Experiment]:
     internal_generator = multimodal(seed, para=True)
     for xp in internal_generator:
         yield xp
+
 
 # pylint: disable=redefined-outer-name
 @registry.register
@@ -717,3 +720,20 @@ def manyobjective_example(seed: Optional[int] = None) -> Iterator[Experiment]:
             for budget in list(range(100, 5901, 400)):
                 for nw in [1, 100]:
                     yield Experiment(mofunc, optim, budget=budget, num_workers=nw, seed=next(seedg))
+
+
+@registry.register
+def far_optimum_es(seed: tp.Optional[int] = None) -> Iterator[Experiment]:
+    # prepare list of parameters to sweep for independent variables
+    seedg = create_seed_generator(seed)
+    popsizes = [5, 40]
+    es = [ng.families.EvolutionStrategy(recombinations=recomb, only_offsprings=False, popsize=pop)
+          for recomb in [0, 1] for pop in popsizes]
+    es += [ng.families.EvolutionStrategy(recombinations=recomb, only_offsprings=only, popsize=pop,
+                                         offsprings=10 if pop == 5 else 60)
+           for only in [True, False] for recomb in [0, 1] for pop in popsizes]
+    optimizers = ["CMA", "TwoPointsDE"] + es  # type: ignore
+    for func in FarOptimumFunction.itercases():
+        for optim in optimizers:
+            for budget in [100, 400, 1000, 4000, 10000]:
+                yield Experiment(func, optim, budget=budget, seed=next(seedg))