+++ nevergrad/optimization/test_optimizerlib.py
@@ -21,6 +21,7 @@ import nevergrad as ng
 from ..common.typetools import ArrayLike
 from ..common import testing
 from . import base
+from . import utils
 from . import optimizerlib as optlib
 from .recaster import FinishedUnderlyingOptimizerWarning
 from .optimizerlib import registry
@@ -81,7 +82,7 @@ def check_optimizer(
         slope, intercept = fitness.get_factors()
         print(f"For your information: slope={slope} and intercept={intercept}")
     # check population queue
-    if hasattr(optimizer, "population"):  # TODO add a PopBasedOptimizer
+    if hasattr(optimizer, "population") and isinstance(optimizer.population, utils.Population):  # type: ignore
         assert len(optimizer.population._queue) == len(set(optimizer.population._queue)), "Queue has duplicated items"  # type: ignore
     # make sure we are correctly tracking the best values
     archive = optimizer.archive
@@ -173,11 +174,8 @@ def test_optimizers_suggest(name: str) -> None:  # pylint: disable=redefined-out
 
 
 # pylint: disable=redefined-outer-name
-@pytest.mark.parametrize("with_parameter", [True, False])  # type: ignore
 @pytest.mark.parametrize("name", [name for name in registry])  # type: ignore
-def test_optimizers_recommendation(with_parameter: bool,
-                                   name: str,
-                                   recomkeeper: RecommendationKeeper) -> None:
+def test_optimizers_recommendation(name: str, recomkeeper: RecommendationKeeper) -> None:
     # set up environment
     optimizer_cls = registry[name]
     if name in UNSEEDABLE:
@@ -188,7 +186,7 @@ def test_optimizers_recommendation(with_parameter: bool,
         random.seed(12)  # may depend on non numpy generator
     # budget=6 by default, larger for special cases needing more
     budget = {"WidePSO": 100, "PSO": 100, "MEDA": 100, "EDA": 100, "MPCEDA": 100, "TBPSA": 100}.get(name, 6)
-    if isinstance(optimizer_cls, optlib.DifferentialEvolution):
+    if isinstance(optimizer_cls, (optlib.DifferentialEvolution, optlib.EvolutionStrategy)):
         budget = 80
     dimension = min(16, max(4, int(np.sqrt(budget))))
     # set up problem
@@ -196,9 +194,7 @@ def test_optimizers_recommendation(with_parameter: bool,
     with warnings.catch_warnings():
         # tests do not need to be efficient
         warnings.filterwarnings("ignore", category=base.InefficientSettingsWarning)
-        param: tp.Union[int, ng.p.Instrumentation] = (dimension if not with_parameter else
-                                                      ng.p.Instrumentation(ng.p.Array(shape=(dimension,))))
-        optim = optimizer_cls(instrumentation=param, budget=budget, num_workers=1)
+        optim = optimizer_cls(instrumentation=dimension, budget=budget, num_workers=1)
         optim.instrumentation.random_state.seed(12)
         np.testing.assert_equal(optim.name, name)
         # the following context manager speeds up BO tests
@@ -380,7 +376,10 @@ def test_parametrization_offset(name: str) -> None:
     if "PSO" in name or "BO" in name:
         raise SkipTest("PSO and BO have large initial variance")
     parametrization = ng.p.Instrumentation(ng.p.Array(init=[1e12, 1e12]))
-    optimizer = registry[name](parametrization, budget=100, num_workers=1)
+    with warnings.catch_warnings():
+        # tests do not need to be efficient
+        warnings.filterwarnings("ignore", category=base.InefficientSettingsWarning)
+        optimizer = registry[name](parametrization, budget=100, num_workers=1)
     for k in range(10 if "BO" not in name else 2):
         candidate = optimizer.ask()
         assert candidate.args[0][0] > 100, f"Candidate value[0] at iteration #{k} is below 100: {candidate.args[0][0]}"